+++ test/IntegrationTests/AspNetTests.cs
@@ -22,7 +22,7 @@ using IntegrationTests.Helpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace IntegrationTests.AspNet;
+namespace IntegrationTests;
 
 public class AspNetTests : TestHelper
 {
