+++ test/IntegrationTests/Helpers/TestSettings.cs
@@ -14,22 +14,20 @@
 // limitations under the License.
 // </copyright>
 
-namespace IntegrationTests.Helpers
+namespace IntegrationTests.Helpers; 
+public class TestSettings
 {
-    public class TestSettings
-    {
-        public TracesSettings TracesSettings { get; set; } = null;
+    public TracesSettings TracesSettings { get; set; } = null;
 
-        public MetricsSettings MetricsSettings { get; set; } = null;
+    public MetricsSettings MetricsSettings { get; set; } = null;
 
-        public string Arguments { get; set; } = null;
+    public string Arguments { get; set; } = null;
 
-        public string PackageVersion { get; set; } = string.Empty;
+    public string PackageVersion { get; set; } = string.Empty;
 
-        public string Framework { get; set; } = string.Empty;
+    public string Framework { get; set; } = string.Empty;
 
-        public int AspNetCorePort { get; set; } = 5000;
+    public int AspNetCorePort { get; set; } = 5000;
 
-        public bool EnableStartupHook { get; set; } = true;
-    }
+    public bool EnableStartupHook { get; set; } = true;
 }
