+++ test/IntegrationTests/MsSqlTests.cs
@@ -1,4 +1,4 @@
-// <copyright file="SqlTests.cs" company="OpenTelemetry Authors">
+// <copyright file="MsSqlTests.cs" company="OpenTelemetry Authors">
 // Copyright The OpenTelemetry Authors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
@@ -22,15 +22,15 @@ using IntegrationTests.Helpers;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace IntegrationTests.SqlClient
+namespace IntegrationTests
 {
-    [Collection(SqlClientCollection.Name)]
-    public class SqlTests : TestHelper
+    [Collection(MsSqlCollection.Name)]
+    public class MsSqlTests : TestHelper
     {
         private const string ServiceName = "TestApplication.SqlClient";
         private readonly SqlClientFixture _sqlClientFixture;
 
-        public SqlTests(ITestOutputHelper output, SqlClientFixture sqlClientFixture)
+        public MsSqlTests(ITestOutputHelper output, SqlClientFixture sqlClientFixture)
             : base("SqlClient", output)
         {
             _sqlClientFixture = sqlClientFixture;
